Finished porting BSP and init. Now debugging is required.
