+++ src/main.cpp
@@ -50,111 +50,13 @@ static  OS_STK        AppTaskStartStk[APP_CFG_TASK_START_STK_SIZE];
 */
 
 static  void  AppTaskStart              (void *p_arg);
-static  void  AppTaskCreate             (void);
-static  void  AppEventCreate            (void);
-extern void  OS_CPU_ExceptHndlr (CPU_INT32U  src_id);
-extern void  BSP_IntHandler       (CPU_INT32U     src_nbr);
-extern void  BSP_IntClr (CPU_INT08U  int_id);
-
-#define  REG_INTCPS_SIR_IRQ      (*(CPU_REG32 *)(SOC_AINTC_REGS + INTC_SIR_IRQ))
-#define  REG_INTCPS_SIR_FIQ      (*(CPU_REG32 *)(SOC_AINTC_REGS + INTC_SIR_FIQ))
-#define  REG_INTCPS_CONTROL      (*(CPU_REG32 *)(SOC_AINTC_REGS + INTC_CONTROL))
-#define  REG_INTCPS_ISR_CLEAR(x) (*(CPU_REG32 *)(SOC_AINTC_REGS + INTC_ISR_CLEAR(x)))
-#define  ACTIVE_IRQ_MASK                                (0x7F)
-#define  ACTIVE_FIQ_MASK                                (0x7F)
-
-static CPU_FNCT_PTR BSP_IntVectTbl[SYS_INT_ID_MAX];
+//static  void  AppTaskCreate             (void);
+//static  void  AppEventCreate            (void);
 
 void init(void)
 {
 	halBspInit();
 }
-
-static  void  OS_BSP_ExceptHandler (CPU_INT08U  except_type)
-{
-    switch (except_type) {
-        case OS_CPU_ARM_EXCEPT_RESET:
-        case OS_CPU_ARM_EXCEPT_UNDEF_INSTR:
-        case OS_CPU_ARM_EXCEPT_SWI:
-        case OS_CPU_ARM_EXCEPT_PREFETCH_ABORT:
-        case OS_CPU_ARM_EXCEPT_DATA_ABORT:
-             while (DEF_TRUE) {
-                 ;
-             }
-    }
-}
-
-void  OS_CPU_ExceptHndlr (CPU_INT32U  src_id)
-{
-    switch (src_id) {
-        case OS_CPU_ARM_EXCEPT_IRQ:
-        case OS_CPU_ARM_EXCEPT_FIQ:
-             BSP_IntHandler((CPU_INT32U)src_id);
-             break;
-
-        case OS_CPU_ARM_EXCEPT_RESET:
-        case OS_CPU_ARM_EXCEPT_UNDEF_INSTR:
-        case OS_CPU_ARM_EXCEPT_SWI:
-        case OS_CPU_ARM_EXCEPT_DATA_ABORT:
-        case OS_CPU_ARM_EXCEPT_PREFETCH_ABORT:
-        case OS_CPU_ARM_EXCEPT_ADDR_ABORT:
-        default:
-             OS_BSP_ExceptHandler((CPU_INT08U)src_id);
-             break;
-    }
-}
-
-void  BSP_IntClr (CPU_INT08U  int_id)
-{
-    CPU_INT08U  reg_nbr;
-
-
-    if (int_id > SYS_INT_ID_MAX) {
-        return;
-    }
-
-    reg_nbr = int_id / 32;
-
-    REG_INTCPS_ISR_CLEAR(reg_nbr) = DEF_BIT(int_id % 32);
-}
-
-void  BSP_IntHandler (CPU_INT32U  src_nbr)
-{
-    CPU_INT32U    int_nbr;
-    CPU_FNCT_PTR  isr;
-
-
-    switch (src_nbr) {
-        case OS_CPU_ARM_EXCEPT_IRQ:
-             int_nbr = REG_INTCPS_SIR_IRQ;
-
-             isr = BSP_IntVectTbl[int_nbr & ACTIVE_IRQ_MASK];
-             if (isr != (CPU_FNCT_PTR)0) {
-                 isr((void *)int_nbr);
-             }
-			 BSP_IntClr(int_nbr);                               /* Clear interrupt									   	*/
-
-			 REG_INTCPS_CONTROL = INTC_CONTROL_NEWIRQAGR;
-             break;
-
-        case OS_CPU_ARM_EXCEPT_FIQ:
-             int_nbr = REG_INTCPS_SIR_FIQ;
-
-             isr = BSP_IntVectTbl[int_nbr & ACTIVE_FIQ_MASK];
-             if (isr != (CPU_FNCT_PTR)0) {
-                 isr((void *)int_nbr);
-             }
-			 BSP_IntClr(int_nbr);                       		/* Clear interrupt									   	*/
-
-			 REG_INTCPS_CONTROL = INTC_CONTROL_NEWFIQAGR;
-             break;
-
-        default:
-             break;
-    }
-}
-
-
 
 static  void  AppTaskStart (void *p_arg)
 {
@@ -167,10 +69,15 @@ static  void  AppTaskStart (void *p_arg)
  #if (OS_TASK_STAT_EN > 0)
     OSStatInit();                                               /* Determine CPU capacity                               */
 #endif
-  
+      //ConsoleUtilsPrintf("\n\n\r");
+      //ConsoleUtilsPrintf("Creating Application Objects...\n\r");
+      //AppEventCreate();                                           /* Create Application Events                            */
+      //ConsoleUtilsPrintf("Creating Application Tasks...\n\r");
+      //AppTaskCreate();                                            /* Create Application Tasks                             */
+    
       while (DEF_TRUE) {
         ConsoleUtilsPrintf("Task 1 message %d!\r\n", i++);
-        OSTimeDlyHMSM(0, 0, 0, 100);
+        OSTimeDlyHMSM(0, 0, 1,0);
     }
 }
 
@@ -199,10 +106,7 @@ int main()
     CPU_ERR  cpu_err;
 #endif
     init();
- 
-    //BSP_BranchPredictorEn();                                    /* Enable Branch Predictor and Cache.                   */
-    //BSP_CachesEn();
-    
+     
     ConsoleUtilsPrintf("Platform initialized.\r\n");
 
     OSInit();                                                   /* Init uC/OS-II                                        */
